+++ src/components/FormLogin.tsx
@@ -7,19 +7,15 @@ import type { LoginResponse, UserLogin } from "../types/user";
 import { useUserStore } from "../store/userStore";
 import { auth } from "../helpers/auth";
 import { toast } from "react-toastify";
-import { authFireBase, provider } from "../firebase/firebase";
-import {
-  getAdditionalUserInfo,
-  signInWithPopup,
-  type UserCredential,
-} from "firebase/auth";
 import { AiFillEye, AiFillEyeInvisible } from "react-icons/ai";
+import LoginGoogle from "./LoginGoogle";
+
+
 
 const FormLogin = () => {
   const [login, setLogin] = useState(false);
   const getUser = useUserStore((state: any) => state.getUser);
   const loginUser = useUserStore((state: any) => state.loginUser);
-  const googleLogin = useUserStore((state: any) => state.googleLogin);
   const [password, setPassword]= useState(true)
 
   useEffect(() => {
@@ -27,46 +23,7 @@ const FormLogin = () => {
     if (isAuth) window.location.href = "/";
   }, []);
 
-  const handleGoogle = async () => {
-    try {
-      const result: UserCredential = await signInWithPopup(
-        authFireBase,
-        provider
-      );
-      // console.log(result);
-      // mando el email a un endpont de back para rectificar su excistencia
-
-      const aditionalinfo = getAdditionalUserInfo(result);
-      // console.log(aditionalinfo);
-
-      const isNewUser = aditionalinfo?.isNewUser;
-      if (isNewUser) {
-        const userGoogleInfo = {
-          email: result.user.email as string,
-          lastName: aditionalinfo?.profile?.family_name as string,
-          name: aditionalinfo?.profile?.given_name as string,
-        };
-        localStorage.setItem("infoGoogle", JSON.stringify(userGoogleInfo));
-
-        // hace hace register con lo valores que no me trae el result.user
-        window.location.href = "/auth/registergoogle";
-      } else {
-        const logGoogle: LoginResponse = await googleLogin(result.user.email);
-        // se hara la logica del login donde se va a guardar todo en zustand
-        // console.log(logGoogle);
-        localStorage.setItem("token", logGoogle.token);
-        getUser(logGoogle.userId).then(() => {
-          setLogin(true);
-          toast.success("Usuario conectado");
-          setTimeout(() => {
-            window.location.href = "/";
-          }, 1000);
-        });
-      }
-    } catch (error) {
-      console.log(error);
-    }
-  };
+  
   const passwordVisibility = () => {
     setPassword(!password);
   };
@@ -130,16 +87,17 @@ const FormLogin = () => {
                     )}
                     </div>
                     </div>
+            
             <Button type="submit" children="Iniciar Sesión" className="mt-6" />
             <p className="text-sm mt-4">
               No tienes cuenta?{" "}
               <a className="font-bold text-primary" href="/auth/register">
                 Registrate
               </a>
-            </p>
-            <Button className="text-xs p-0.5 w-fit" onClick={handleGoogle}>
-              Iniciar sesión con google
-            </Button>
+              <div className="mt-4 flex justify-center items-center">
+              <LoginGoogle/>
+              </div>
+            </p>          
           </Form>
         )}
       </Formik>
