+++ src/pages/new-shipment.astro
@@ -4,5 +4,5 @@ import Shipment from '../components/Shipment';
 ---
 
 <Layout title="Hacer un envio">
-    <Shipment client:load/>
+    <Shipment client:load transition:persist/>
 </Layout>
